+++ lib/appmap/rspec.rb
@@ -13,58 +13,9 @@ module AppMap
       AppMap.detect_metadata
     end
 
-    module FeatureAnnotations
-      def feature
-        return nil unless annotations
-
-        annotations[:feature]
-      end
-
-      def labels
-        labels = metadata[:appmap]
-        if labels.is_a?(Array)
-          labels
-        elsif labels.is_a?(String) || labels.is_a?(Symbol)
-          [ labels ]
-        else
-          []
-        end
-      end
-
-      def feature_group
-        return nil unless annotations
-
-        annotations[:feature_group]
-      end
-
-      def annotations
-        metadata.tap do |md|
-          description_args_hashes.each do |h|
-            md.merge! h
-          end
-        end
-      end
-
-      protected
-
-      def metadata
-        return {} unless example_obj.respond_to?(:metadata)
-
-        example_obj.metadata
-      end
-
-      def description_args_hashes
-        return [] unless example_obj.respond_to?(:metadata)
-
-        (example_obj.metadata[:description_args] || []).select { |arg| arg.is_a?(Hash) }
-      end
-    end
-
     # ScopeExample and ScopeExampleGroup is a way to handle the weird way that RSpec
     # stores the nested example names.
     ScopeExample = Struct.new(:example) do
-      include FeatureAnnotations
-
       alias_method :example_obj, :example
 
       def description?
@@ -83,8 +34,6 @@ module AppMap
     # As you can see here, the way that RSpec stores the example description and
     # represents the example group hierarchy is pretty weird.
     ScopeExampleGroup = Struct.new(:example_group) do
-      include FeatureAnnotations
-
       alias_method :example_obj, :example_group
 
       def description_args
@@ -133,9 +82,15 @@ module AppMap
           page.driver.options[:http_client].instance_variable_get('@server_url').port
         end
 
-        warn "Starting recording of example #{example}" if AppMap::RSpec::LOG
+        warn "Starting recording of example #{example}@#{source_location}" if AppMap::RSpec::LOG
         @trace = AppMap.tracing.trace
         @webdriver_port = webdriver_port.()
+      end
+
+      def source_location
+        result = example.location_rerun_argument.split(':')[0]
+        result = result[2..-1] if result.index('./') == 0
+        result
       end
 
       def finish
@@ -148,22 +103,16 @@ module AppMap
 
         AppMap::RSpec.add_event_methods @trace.event_methods
 
-        class_map = AppMap.class_map(@trace.event_methods, include_source: false)
+        class_map = AppMap.class_map(@trace.event_methods, include_source: AppMap.include_source?)
 
         description = []
         scope = ScopeExample.new(example)
-        feature_group = feature = nil
 
-        labels = []
         while scope
-          labels += scope.labels
           description << scope.description
-          feature ||= scope.feature
-          feature_group ||= scope.feature_group
           scope = scope.parent
         end
 
-        labels = labels.map(&:to_s).map(&:strip).reject(&:blank?).map(&:downcase).uniq
         description.reject!(&:nil?).reject!(&:blank?)
         default_description = description.last
         description.reverse!
@@ -177,24 +126,10 @@ module AppMap
 
         full_description = normalize.call(description.join(' '))
 
-        compute_feature_name = lambda do
-          return 'unknown' if description.empty?
-
-          feature_description = description.dup
-          num_tokens = [2, feature_description.length - 1].min
-          feature_description[0...num_tokens].map(&:strip).join(' ')
-        end
-
-        feature_group ||= normalize.call(default_description).underscore.gsub('/', '_').humanize
-        feature_name = feature || compute_feature_name.call if feature_group
-        feature_name = normalize.call(feature_name) if feature_name
-
         AppMap::RSpec.save full_description,
                            class_map,
-                           events: events,
-                           feature_name: feature_name,
-                           feature_group_name: feature_group,
-                           labels: labels.blank? ? nil : labels
+                           source_location,
+                           events: events
       end
     end
 
@@ -227,12 +162,11 @@ module AppMap
         @event_methods += event_methods
       end
 
-      def save(example_name, class_map, events: nil, feature_name: nil, feature_group_name: nil, labels: nil)
+      def save(example_name, class_map, source_location, events: nil, labels: nil)
         metadata = AppMap::RSpec.metadata.tap do |m|
           m[:name] = example_name
+          m[:source_location] = source_location
           m[:app] = AppMap.configuration.name
-          m[:feature] = feature_name if feature_name
-          m[:feature_group] = feature_group_name if feature_group_name
           m[:labels] = labels if labels
           m[:frameworks] ||= []
           m[:frameworks] << {