+++ test/rspec_test.rb
@@ -28,8 +28,6 @@ class RSpecTest < Minitest::Test
       assert_includes appmap.keys, 'metadata'
       metadata = appmap['metadata']
       assert_equal 'Inventory', metadata['name']
-      assert_includes metadata.keys, 'labels'
-      assert_equal metadata['labels'], %w[inventory]
     end
   end
 
@@ -42,9 +40,8 @@ class RSpecTest < Minitest::Test
       assert_equal AppMap::APPMAP_FORMAT_VERSION, appmap['version']
       assert_includes appmap.keys, 'metadata'
       metadata = appmap['metadata']
-      assert_equal 'Saying hello', metadata['feature_group']
-      assert_equal 'Speak hello', metadata['feature']
       assert_equal 'Hello says hello', metadata['name']
+      assert_equal 'spec/decorated_hello_spec.rb', metadata['source_location']
       assert_includes metadata.keys, 'client'
       assert_equal({ name: 'appmap', url: AppMap::URL, version: AppMap::VERSION }.stringify_keys, metadata['client'])
       assert_includes metadata.keys, 'recorder'
@@ -63,8 +60,6 @@ class RSpecTest < Minitest::Test
       appmap = JSON.parse(File.read(appmap_file))
       assert_includes appmap.keys, 'metadata'
       metadata = appmap['metadata']
-      assert_equal 'Hello', metadata['feature_group']
-      assert_equal 'Hello', metadata['feature']
       assert_equal 'Hello says hello', metadata['name']
     end
   end
@@ -76,7 +71,6 @@ class RSpecTest < Minitest::Test
       appmap = JSON.parse(File.read(appmap_file))
       assert_includes appmap.keys, 'metadata'
       metadata = appmap['metadata']
-      assert_equal %w[hello speak], metadata['labels'].sort
     end
   end
 end